Update chessBoard
Last steps to do are create cases for draws, en passant and to add in ability to tell if checkmated. Current plan for checkmate mechanic is to use kingCheck(opp) then via not move() instead use board[][] = board[][] to directly make allPosMoves(opp) followed by kingCheck to see if it removes kingCheck(), and if it doesn't to declare checkmate via gameOver() valid = true.